Removed wallet.json. Added error handling in the case of incorrect input. Display transaction.json after launch
