+++ src/network/peer.rs
@@ -14,9 +14,9 @@ use crate::utils::hash::hash_as_string;
 use crate::utils::save_and_load::load_object;
 use crate::utils::save_and_load::save_object;
 use crate::utils::sign_and_verify;
+use crate::utils::sign_and_verify::PrivateKey;
 use crate::utils::sign_and_verify::PublicKey;
 use crate::utils::sign_and_verify::Verifier;
-use crate::wallet::Wallet;
 use ed25519_dalek::Keypair;
 use local_ip_address::local_ip;
 use log::{error, info, warn};
@@ -25,7 +25,6 @@ use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 use std::collections::HashMap;
 use std::collections::HashSet;
-use std::io::Read;
 use std::io::Write;
 use std::{env, fs, io};
 use std::{fs::File, path::Path};
@@ -210,7 +209,7 @@ impl Peer {
 
             // We create a new wallet for each peer
             let (private_key_initial, public_key_initial) = sign_and_verify::create_keypair();
-            let values = vec![(
+            let wallet: Vec<(PrivateKey, PublicKey, Outpoint, u32)> = vec![(
                 private_key_initial,
                 public_key_initial,
                 Outpoint {
@@ -219,7 +218,6 @@ impl Peer {
                 },
                 500,
             )];
-            let wallet: Wallet = Wallet(values);
             save_object(&wallet, String::from("wallet"), String::from("system"));
         }
 
@@ -276,44 +274,6 @@ impl Peer {
             let port = p.clone();
             let tx_clone = tx.clone();
             tokio::spawn(async move { Peer::listen(ip, port, tx_clone).await });
-        }
-
-        // Give the peer a chance to see what a transaction.json file looks like
-        info!(
-            "In this system you can include your own transaction.json file under the 'account' folder in the root. Would you like to see what
-            a template transaction.json file looks like? y/n"
-        );
-        let mut choice_display: String = String::new();
-        io::stdin()
-            .read_line(&mut choice_display)
-            .expect("Failed to read line");
-
-        if choice_display.to_lowercase().trim() == "y" {
-            let dirname = String::from("account");
-            let object_name = String::from("transaction");
-
-            let example_transaction = get_example_transaction();
-            save_object(
-                &example_transaction,
-                String::from("transaction"),
-                String::from("account"),
-            );
-
-            let slash = if env::consts::OS == "windows" {
-                "\\"
-            } else {
-                "/"
-            };
-            let mut file =
-                File::open(dirname + slash + &object_name + ".json").expect("Failed to open file");
-            let mut contents = String::new();
-            file.read_to_string(&mut contents)
-                .expect("Failed to read file");
-
-            let json_data: serde_json::Value =
-                serde_json::from_str(&contents).expect("Failed to parse JSON");
-
-            println!("{}", serde_json::to_string_pretty(&json_data).unwrap());
         }
 
         return tx.clone();
